// src/common-pages/proxy/proxy.type.ts
export interface IProxyUserInfo {
  activeDay: string;
  agentLevel: number;
  betAmount: string;
  betUserCount: number;
  commissionAmount: string;
  newUserCount: number;
  rechargeAmount: string;
  registerDate: string;
  totalUserCount: number;
  userAvatar: string;
  userId: number;
  userPhone: string;
  wonLostAmount: string;
  contactPhone: string;
  rechargeUserCount: number;
}